refactor(collector): fold gRPC support into client
Keep the collector wire codec and Tonic service generation with the client, then re-export CollectorServer from the server crate. Remove the standalone shared RPC crate and its direct consumers.
