+++ Cargo.toml
@@ -11,7 +11,6 @@ members = [
     "crates/codegen",
     "crates/collector/server",
     "crates/collector/client",
-    "crates/collector/rpc",
     "crates/events",
     "crates/io",
     "crates/io/callback",
@@ -89,7 +88,6 @@ default-members = [
     "crates/codegen",
     "crates/collector/server",
     "crates/collector/client",
-    "crates/collector/rpc",
     "crates/events",
     "crates/io",
     "crates/io/callback",
