+++ crates/collector/client/Cargo.toml
@@ -8,7 +8,6 @@ publish.workspace = true
 bitcode.workspace = true
 bytes.workspace = true
 http.workspace = true
-quent-collector-rpc = { path = "../rpc" }
 quent-events = { workspace = true, features = ["serde"] }
 serde.workspace = true
 thiserror.workspace = true
@@ -18,3 +17,6 @@ tokio-util.workspace = true
 tonic.workspace = true
 tracing.workspace = true
 uuid.workspace = true
+
+[build-dependencies]
+tonic-build.workspace = true
