--- crates/collector/rpc/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "quent-collector-rpc"
-version.workspace = true
-edition.workspace = true
-publish.workspace = true
-
-[dependencies]
-bytes.workspace = true
-tonic.workspace = true
-
-[build-dependencies]
-tonic-build.workspace = true
