+++ crates/collector/server/Cargo.toml
@@ -6,7 +6,6 @@ version.workspace = true
 
 [dependencies]
 quent-collector-client = { path = "../client" }
-quent-collector-rpc = { path = "../rpc" }
 tokio = { workspace = true, features = ["rt-multi-thread"] }
 tokio-stream.workspace = true
 tonic.workspace = true
