+++ crates/collector/server/src/lib.rs
@@ -7,6 +7,8 @@
 
 pub mod server;
 
+pub use quent_collector_client::grpc::collector_server::CollectorServer;
+
 /// Re-exported so callers constructing a [`server::CollectorService`] over a
 /// generic context type can bound it, and decode ingested events, without
 /// depending on `quent-collector-client` directly.
