+++ crates/instrumentation/Cargo.toml
@@ -42,7 +42,6 @@ criterion = { version = "0.8", features = ["html_reports"] }
 http.workspace = true
 pprof = { version = "0.15", features = ["flamegraph"] }
 quent-collector = { path = "../collector/server" }
-quent-collector-rpc = { path = "../collector/rpc" }
 quent-events = { path = "../events", features = ["serde"] }
 quent-io = { path = "../io", features = ["ndjson"] }
 serde = { workspace = true, features = ["derive"] }
